# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Organizations",
        "description": "Documentación de la API de prueba técnica",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar usuario",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "David"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "david@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login usuario",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "david@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso"
                    },
                    "401": {
                        "description": "Credenciales incorrectas"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Usuario autenticado",
                "operationId": "9d9bc556f927575aa492a88c28fe5908",
                "responses": {
                    "200": {
                        "description": "Usuario autenticado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cerrar sesión",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logout exitoso"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/organizations/{id}/invitations": {
            "post": {
                "tags": [
                    "Invitations"
                ],
                "summary": "Crear invitación",
                "operationId": "2d28d14773ba81aba5d4dc132c4c8a23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la organización",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "role"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "manager@test.com"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "manager"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Invitación creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invitación creada correctamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "manager@test.com"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "manager"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "uuid-token"
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "example": "2026-03-21 12:00:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "Sin permisos"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/invitations/{token}": {
            "get": {
                "tags": [
                    "Invitations"
                ],
                "summary": "Consultar invitación por token",
                "operationId": "fd4e19c9b9bc9b279113967de78349f8",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Token de la invitación",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "uuid-token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invitación encontrada"
                    },
                    "400": {
                        "description": "Invitación expirada o usada"
                    },
                    "404": {
                        "description": "No encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/invitations/{token}/accept": {
            "post": {
                "tags": [
                    "Invitations"
                ],
                "summary": "Aceptar invitación",
                "operationId": "8bda439589b1e62700acd5cf2908adf5",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Token de la invitación",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "uuid-token"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "David"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invitación aceptada"
                    },
                    "400": {
                        "description": "Error de validación o invitación usada"
                    },
                    "404": {
                        "description": "Invitación no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/organizations": {
            "post": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Crear organización",
                "operationId": "2d849ce559213546c5ec62ee8ea255db",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Empresa Tech"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Organización creada"
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "Sin permisos"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Invitation": {
                "title": "Invitation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "email": {
                        "type": "string",
                        "example": "manager@test.com"
                    },
                    "organization_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "role": {
                        "type": "string",
                        "example": "manager"
                    },
                    "token": {
                        "type": "string",
                        "example": "uuid-token"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    }
                },
                "type": "object"
            },
            "Organization": {
                "title": "Organization",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Empresa Tech"
                    },
                    "owner_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "David"
                    },
                    "email": {
                        "type": "string",
                        "example": "david@test.com"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Invitations",
            "description": "Invitations"
        },
        {
            "name": "Organizations",
            "description": "Organizations"
        }
    ]
}